move dynamic fields helpers together
