+++ lib/frontend_helpers.rb
@@ -7,13 +7,14 @@ require 'frontend_helpers/builder_html_utils'
 require 'frontend_helpers/shared_form_builder_utils'
 require 'frontend_helpers/field_group_wrapper'
 require 'frontend_helpers/field_group_builders'
+require 'frontend_helpers/form_builder_helpers/dynamic_fields'
 require 'frontend_helpers/form_builder'
 require 'frontend_helpers/layout_concern'
-require 'frontend_helpers/dynamic_fields_helper'
 require 'frontend_helpers/components'
 require 'frontend_helpers/custom_redirect_concern'
 
 module FrontendHelpers
   class Engine < ::Rails::Engine; end
+
   class Error < StandardError; end
 end
